add category update job

+++ modules/quhao.appserver/app/com/withiter/models/merchant/CategoryEntityDef.java
@@ -5,7 +5,7 @@ import com.withiter.models.BaseModel;
 
 public abstract class CategoryEntityDef extends BaseModel {
 
-	public int count = 0;
+	public long count = 0;
 	public String cateType;
 	
 }
